Use the VERSION variable instead of recalculating it
This matters for calculating the Python include directory.
In CPython it is always
"include/python%d.%d" % sys.version_info[:2]

but alternative implementations may have different version strings
in order to not conflict with CPython. For example the version
string for Pyston is "3.8-pyston2.3" so that the include directory
is "include/python3.8-pyston2.3".

The usage of VERSION seems somewhat canonical; for
example see build_scripts.py:104